Refactor cli opening it up for more subcommands
