+++ docs/fides/mkdocs.yml
@@ -30,9 +30,9 @@ nav:
       - Write a Policy: tutorial/policy.md
       - Add Google Analytics: tutorial/google.md
       - Manage Google Analytics with Fidesctl: tutorial/pass.md
-  - How-To Guides:
-      - Generating a Dataset: guides/generate_dataset.md
-      - Scanning a Resource: guides/scan_resource.md
+  - How-To Guides: 
+      - Generating Resources: guides/generate_resource.md
+      - Scanning Resources: guides/scan_resource.md
       - Example Policies: guides/policies.md
   - CI/CD Reference: ci_reference.md
   - Fides Language:
